Use mockup to to implement RTF output.
A lamprop text output file was formatted in MS-Word, and then repeatedly
edited in a text editor to make it simpler and then and opened in MS-Wordpad
to verify that it still worked. This has resulted in an RTF file that was used
as a basis for this RTF output module.