Change output API for easier testing. Remove RTF output.
Output functions now return a list of strings. This makes it easier for
the tests to compare output with that of a previous version.

Output as lines does not fit well with RTF output, so RTF is hereby
deprecated and removed. Since word processors can import HTML just fine,
using RTF was not really neseccasy.